requirements.txt
coloredlogs==14.0
colorgram.py==1.2.0
